requirements.txt
streamlit>=1.40.0
pandas>=2.2.0
plotly==5.24.0
pyyaml>=6.0
openpyxl>=3.1.0
numpy>=2.1.0
pytest>=8.0.0
